refactor: consolidate duplicated ModuleCategory and type helpers into shared types module
